tested. moved code into folders to clean up
