+++ db.js
@@ -1,10 +1,10 @@
 const Pool = require("pg").Pool;
 const pool = new Pool({
-  user: "postgres",
-  password: "Katy3568",
-  host: "localhost",
-  port: 5432,
-  database: "zotnfound",
+  user: process.env.PG_USER,
+  password: process.env.PG_PASSWORD,
+  host: process.env.PG_HOST,
+  port: process.env.PG_PORT,
+  database: process.env.PG_DB_NAME,
 });
 
 module.exports = pool;
