+++ package.json
@@ -4,7 +4,8 @@
   "description": "zotnfound backend",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon server.js"
   },
   "author": "Steven Zhou",
   "license": "ISC",
